refactor: create forms directory and move config files
