--- packages/rockets-next-web/src/app/(main)/users/types.ts
@@ -1,1 +1,0 @@
-export type ActionType = "creation" | "edit" | "details" | "delete" | null;
